@@ -3,18 +3,18 @@
   "version": "0.1.0",
   "private": true,
   "dependencies": {
-    "ajv": "^6.12.6",
-    "cross-env": "^5.2.1",
+    "ajv": "^8.12.0",
+    "cross-env": "^7.0.3",
     "file-saver": "^2.0.5",
-    "flow-bin": "^0.80.0",
-    "memoize-one": "^3.1.1",
-    "npm": "^9.7.2",
-    "react": "^16.14.0",
-    "react-dom": "^16.14.0",
-    "react-redux": "^5.1.2",
-    "react-scripts": "3.4.4",
-    "redux": "^4.0.5",
-    "redux-thunk": "^2.3.0"
+    "flow-bin": "^0.208.0",
+    "memoize-one": "^6.0.0",
+    "npm": "^9.8.0",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
+    "react-redux": "^8.1.1",
+    "react-scripts": "5.0.1",
+    "redux": "^4.2.1",
+    "redux-thunk": "^2.4.2"
   },
   "scripts": {
     "start": "cross-env HTTPS=true react-scripts start",
@@ -35,5 +35,8 @@
       "last 1 firefox version",
       "last 1 safari version"
     ]
+  },
+  "devDependencies": {
+    "@testing-library/react": "^14.3.1"
   }
 }
